Fetch domain xml from libvirt for virt-v2v-in-place
Rather than constructing our own libvirt xml to pass to virt-v2v, we
should be using libvirt itself to connect to the vsphere server and
generate a domain xml file for the domain that we are trying to migrate.
This ensures that virt-v2v has as complete information as possible to
perform conversion of the domain.

This commit adds a libvirt dependency to the virt-v2v container that is
used to query libvirt for the domain xml description. We no longer need
to generate pseudo-libvirtxml in the controller and pass it down to the
conversion pod via ConfigMap. But we do need to pass the host secret
down to the conversion pod in the same way that we do when we're not
converting in-place since we now need to connect to the source provider
(via libvirt) for this use case.

Resolves: MTV-2799

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>